Initial Commit For CRUD-Harvest user story
Created a new Page for Harvest in React. Setup a controller, a HarvestDAO, and a HarvestDAOStub whom both follow an interface, IHarvestDAO, and uses the Harvest Model class.

Created a HarvestServices.tsx in React which is used to collect data from the API.

The data stored in HarvestDAOStub is successfully retrieved and displayed on the Harvest Page.